Fix minor warnings
All warnings that don’t matter much, but are annoying:
- Make conversions between enums explicit
- uints can’t be smaller than 0.
- The result of vector::size should be stored in a size_t variable,
especially if you then want to cast it to a pointer. Anything else is
wrong in an LP64 model (Mac) or LLP64 model (Windows) (it would be okay
in an ILP64 model, which I think is prevalent on Linux).
- Clang’s control flow analysis can’t determine that the switch
statement covers all cases even though the if makes sure that other
values are impossible. A default case that would abort the program
fixes that.