Setup Lua for Mac OS X
Lua worked, but due to a lack of defines, it didn’t use all available
Mac OS X options (for example it uses the insecure tmpfile, which was a
warning in OpenTomb). This was fixed. Also apparently it needs
libreadline.dylib for something, maybe, so that’s linked in too.