Turn bordered texture atlas into proper C++ class
I’ve been wanting to do this for a while. This style is just a lot
cleaner. Ideally I’d be able to use std::vector for the things that
need it, but apparently that’s not popular at the moment, so the old
array handling code remains for now.

Apart from being a class, other parts were also C++ified: References
for things that cannot legally be null, and bool instead of int to
store conditions.