Resource cleanup: Generating faces
I hate resource.cpp. I love the rest of OpenTomb (more or less), but I
hate this file with a burning passion. As long as the Odyssey, and
about as hopeful, all just the same code repeated over and over and
over again to reach over 4000 lines. It must be punished, by cutting
out as much code as possible, so one can finally see what is going and
why.

This commit focuses on generating faces from meshes and rooms, which
had a lot of duplication that could very easily be refactored.