Use standard C++11 without extensions
The default for Xcode is to use a version it calls GNU++11, which is
C++11 with some GCC specific extensions (specifically ones that Clang
copied; Xcode uses Clang unless told otherwise). I don’t want to risk
breaking the Windows build, so I’ll disable it and not use any of the
extensions unless the Windows version starts doing so first.