Extend climb flags upwards
A common pattern in TR levels is to set climb flags only on the bottom room, but expect them to apply all the way up through several rooms until hitting a ceiling with no following room. This is for example seen in the wooden house at the start of "Venice"/BOAT.TR2: Climb flag is set in the bottom room. The room above does not have it, and consequently, the ladder could not be climbed in OpenTomb. This commit copies the climbing flags upwards, which solves this problem.