Make wall climbing work reliably
The previous fix for wall climbing issues only worked in one direction
and missed a lot of edge cases. By moving the flag propagation out of
the resource loading and into the actual query, this solves that issue.

This also fixes the part where I suspected that there was something
wrong with ladders coming out of water. No, the two instances I tried
were just examples where the room order was „wrong“.