Generalized sector_above and sector_below calculations
This finally fixes a bug which prevented Lara to activate triggers below
her current level, when something was blocking her from reaching it
(e.g. first The Great Wall trapdoor).
Another change is introducing gameflow opcode queue, which allows to add
several gameflow operations in the same game loop. While it's not useful
for now, it could be handy when more gameflow opcodes are implemented.
Also, this commit forces HP bar to be drawn when weapons are armed.