Get from wade to swim properly
The transition from wading to swimming happened only when Lara had
waded completely underwater. This adds an offset ensuring that she
starts wading about at the same time as she does in the real game. Test
case was „The Great Wall“. Values and maths here is largely guesswork
so far.