Add error reporting to all lua calls
And I mean all of them. There are a lot, aren’t there? Maybe I’ll later
convert this to some utility function so this can be done in one call,
and also perhaps include the line number, function and file where the
call was made, but this should provide an advantage already.