Complete generalized lua error handling
The new function (actually a macro) lua_CallAndLog is meant as a
replacement for lua_pcall. It calls lua_pcall internally, and if an
error occurs, logs this error, together with the C location where the
code was called. This should make it a lot easier to figure out what is
going wrong in Lua code.