Matrix cleanup [1/2]
This gets rid of all the glRotate, glScale and glTranslate calls in
OpenTomb, except for one that is commented out. This is just a
straightforward translation of the existing code, which may not always
be optimal (see gui.cpp:741: Why is there a cos when trying to
calculate an angle? That’s almost certainly wrong). Note that this is
not yet enough to easily switch to OpenGL ES/Core Profile; for that we
have to get rid of all the glMultMatrix calls as well, and as a
corollary, glPushMatrix and glPopMatrix.