Deal with zero x camera offsets properly
And also make the x camera offset nonzero again, because that seems
like debug code that should not be in the repository.

A dx of zero generates a zero-length ray for the collision detection of
the camera, which is illegal and triggers an exception (implemented as
an explicit breakpoint, which is rare) deep in bullet. With this
commit, no collision detection is performed in this case.