Fix transparent polygons not showing up
This fix is unnecessary. The previous commit already fixed that issue.
But it did so accidentally, by changing global state that just happened
to be used by the following stage as well. I dislike how non-obvious
that is, so I set the state here again so it’s clear where what is
coming from.