Implement proper light radius calculation.
This was related to #46, but sadly does not fix it. Light calculations do work a lot better now, for example in setgamef(4,3). However, the code to parse TR5 light intensities is seemingly rather broken. The current implementation sanitizes it somewhat, but the result is that TR5 lights just don't show up, period. Only ambient light is used.