Add initial support for VAOs
Vertex Array Objects reduce the required setup for draw calls
drastically. Sadly, they’re not supported everywhere, so a wrapper
around them was created. This wrapper is somewhat limited; it supports
only fully custom vertex attributes. This is deliberate because that’s
a requirement for OpenGL 3.2 Core Profile. The wrapper will use normal
setup calls if necessary, or vertex array objects if the required
extension (either in its original APPLE or its approved ARB version) is
present. I guess it should check for the OpenGL version as well, to be
thorough, but practically I think this should cover most bases.

Currently only used for room sprites, because those were the easiest to
convert. More to come.