Fix seams in animated textures
Due to wrong iterating over the animated polygons, seams appeared in
animated textures. This had no issue of its own but was mentioned in
#58.

Note that textures in TR3 still have somewhat noticeable seams. After
lots of experimenting and looking at screenshots, I’ve concluded that
these are caused by the original textures and thus are not a bug but a
feature. If you disagree, please open a new ticket. Bonus points if you
have a good screenshot of what it should look like!