Change skinning to use matrix indices, not weights
With this change, it is possible to use many more matrices, which will
be relevant for hair later (#51). By keeping the weights fixed at 0.5,
we can avoid the need for another vertex attribute. Note that in a more
complex scenario, we would probably need two.

Sadly specifying an ivec2 input isn’t possible in vanilla OpenGL 2.0,
which remains the target, so we’re doing this weird vec2 cast to int
dance here.

Fun fact that I realized while doing this: There is no actual need to
keep the weights normalized; removing the 0.5s in that shader works
just as well - as would any other number > 0. This actually makes sense
if you think about it for a few seconds (hint: How do homogenous
coordinates work again?), but I’m currently reluctant to make use of
that because it is confusing.

Also removed shaders from the copy-to-bundle phase of the Mac build,
because they remain in the OpenTomb folder.