Fix hair rendering.
So this one may seem odd because hair rendering didn't look all that broken (except for the gap between Lara's head and the hair in TR4, which still remains). But it was. If you looked at it with r_coll, you saw that the hair didn't match the models, it rather interpolated between them. That was sort of okay for hair because you didn't notice it (though it did feel unpleasant), but it's not okay if we want to generalize the single-draw approach to all animated meshes, and I think we want to. I at least do.

There's some weird maths going on in render.cpp, so I left my comment explaining what it all meant. Not sure it helps, but there you go.