Change text rendering to use VAO.
Alright! More than a hundred new lines of code just to get no visible change AT ALL! Cool, huh?

Yeah, it seems crazy to me, too. But the truth of the matter is that we want OpenTomb to be a modern engine based on modern code and technologies. That means keeping current best practices, and being open to adopting new things in the future at some point. And where OpenGL is concerned, that means adopting OpenGL 3.2 Core Profile (or OpenGL ES 2.0 with VAOs, which is more or less a subset and what we're currently working towards). And at times, that means writing code that is longer to work around the fact that this feature isn't in OpenGL core anymore, because the OpenGL guys said, "you need OpenGL for the stuff you can't do on your own; anything you can do yourself is something you can probably do better than if we're guessing". Such is life.

Currently missing: Console separating line. I have half a mind to just screw it and use Render_Rect for that. Would save me a considerable amount of trouble.