Get rid of unnecessary state changes.
With everything on VAO and shaders, a lot of state changes are no longer necessary, in particular anything to do with vertex arrays. So let's get rid of them! This commit may break things, but I hope it doesn't. If it does break anything, the result will most likely be a crash somewhere in rendering code. But I don't really expect anything like that.

With that, we're ready to switch to OpenGL 3.2 Core Profile. It will still take a bit of work, but not a lot anymore. Which raises the interesting question of "do we want to?", since it means anyone with a DirectX 9-only card will have a problem. I'll produce a test version soonish, and based on that we can decide.