Add rough spiky ceiling functionality.
But we REALLY need ghost collision callbacks, as currently spike ceiling
may push Lara out of level geometry. Additionally, we get major
slowdowns when Lara is in hanging/climbing state, and spiky ceiling is
approaching - I guess we mush set collision mode to bounding box to fix
that (which is currently not functional).