Get Mac version working again.
This fixes… something in the 11x range. Gotta look it up when I have
internet connection again.

The way it does so is rather hack, which is why this is in a new
branch. First, it adds the custom OpenAL version back in and (unlike
custom bullet and freetype) actually compiles it, because the system
version of OpenAL does not include all the effects stuff.

As part of that, the audio.cpp file assumes that alExt prototypes are
always available. Just roll with it.

I have zero clue what’s going on with the None in luahelper, but the
initializing expression is definitely not constexpr, so I changed it to
static so it gets initialized sooner or later anyway. The whole file is
ugly. Int8 as an optional int8? That’s just begging for trouble.