Change shaders to GLSL 1.5.0
The existing code claimed they were GLSL 4.1.0 but didn’t actually
conform to that. This is fixed now.

(For reference: The GLSL version for OpenGL 3.2 is 1.5.0. Then they
decided that the versioning is weird, so from OpenGL 3.3 on, it’s the
OpenGL’s version number plus a 0 at the end.)