Remove deprecated OpenGL calls
This is necessary as part of OpenGL 3.2 Core Profile compatibility. And
it causes a number of new problems: Rendering in wireframe mode is no
longer possible (eh, who cares); line widths are not set (this could be
annoying), push and pops of graphics state are unsupported (this can
cause all sorts of fun issues) and alpha testing is removed (this is
the big one, as it destroys all transparency in TR1 and 2). Fun stuff.