Make shader stages temporary objects again
Apparently there was some misconception here: It’s perfectly fine to
call glDeleteShader on a shader you’re still using after it has been
attached to a program. If you do that, it will only actually be deleted
once the program is deleted (or the shader gets detached from the
program, but there is absolutely no good reason to do that, ever).