Fix adjacent rooms not being visible through portals
I noticed that some portals had no dest_room set. This caused
Frustum::portalFrustumIntersect to also return a nullptr.

The reason for this is that portal->dest_room is now assigned the
pointer (actually, a shared_pointer) directly. Before, however, it was
assigned a pointer to the world->rooms array.
Combine that with the fact that each room is allocated just before
TR_GenRoom() is entered, so the rooms array is initially filled with
nulls, and you end up assigning at least some nulls to portal->dest_room
as well.

The easiest fix was to initialize all the pointers in the world->rooms
array before calling TR_GenRoom().

The last step was to find, by trial and error, that by switching the
indices into the array of vertices to the state before the C++
refactoring makes the adjacent rooms visible again.