Forbid making copies of Camera
The constructor for Camera gives a frustum object, created inside the
body of the constructor, a pointer to the current camera position,
like so:

    frustum->cam_pos = &m_pos

If a copy copy constructor is called, only a shallow copy is performed,
which means that the pointer in frustum will point to invalid memory if
the object whose values were copied was a temporary.
Whether a copy constructor is called or not is, of course, dependent on
the optimizations performed by the compiler.

The line initializing engine_camera can simply be deleted - variables
with static storage duration are initialized automatically.

Also delete the copy constructor and the copy assignment operator (the
move constructor/assignment operator will be deleted as well) to
make attempts to copy the object fail.