Convert everything sound-related to float.
It's a bit silly to use OpenGL 3.2 and still stick to 16-bit sound. But
it's not the main reason why everything is in float now. The problem is,
libsndfile produces nasty clicks when floating-point data is read into
16-bit buffer (e.g. FLAC and OGG). It has a fix for this problem, but it
involves full scan of sample data on file load, which significantly
slows down playback. So, the better option is to convert everything to
float, as it was with our old custom SDL-backended OpenAL.

Also added new helpers module, where we can finally place some handy
functions to fix some values - like clamping particular variable or
fixing angles (therefore fixes #173).