Replace all Bullet include subdirectories
There is no point to add bullet includes to compiler search path
and at the same time use an absolute path. It's just a remnant
from the days when Bullet was a subdirectory of a project. Bullet
itself uses a different hierarchy. This makes it impossible to
properly include Bullet as both in-tree or as an installed library.