Use custom find_module for SDL2_image instead of PkgConfig.
PkgConfig seems to sort of find SDL2_image, but not the correct path to the library on
Mac OS. The custom find module solves this issue. It also has the side-effect of
removing the reliance on pkg-config as an external tool. That may not be important at
all (I never had a problem with pkg-config not being installed, even on Mac OS), but
it's still nice.