Add the headers of our files to CMakeLists
The header files do not need to be compiled, so it may not seem like they need to be in CMakeLists.txt, but
that is wrong. Changes to them still trigger recompiles elsewhere, so CMake needs to be aware of them to
include them in the generated dependency tracking. Also, without this, the headers aren't in the generated IDE
files, which makes life surprisingly annoying.

(cherry picked from commit 3d819f1041deda82ae52b78bda6fa3ed277f6eab)